// src/pages/interviewEdits/data.d.ts
export interface interviewDetailData{
    interviewId?: string;
    interviewTimes?: number;
    interviewStudyAbility?: number;
    interviewChatAbility?: number;
    interviewSumAbility?: number;
    interviewDigAbility?: number;
    interviewUserId?: string | number;
    interviewAdminName?: string;
    interviewUserName?: string;
    interviewIsStay?: number;
    interviewFuture?: number;
    interviewEvaluation?: string;
}